% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_plots.R
\name{plot_hobos}
\alias{plot_hobos}
\title{Plot HOBO Logger Data}
\usage{
plot_hobos(df)
}
\arguments{
\item{df}{A data frame containing at least the following columns:
\describe{
  \item{datetime}{POSIXct timestamp}
  \item{abspress}{Absolute pressure values}
  \item{temp}{Temperature values}
  \item{dataset}{Dataset/site name}
}}
}
\value{
A ggplot object showing faceted time series of pressure and temperature.
}
\description{
Creates time series plots of absolute pressure and temperature
from one or more HOBO logger datasets. If multiple datasets are provided,
they are overlaid and colored by dataset name.
}
\examples{
\dontrun{
combined <- combine_hobos(Ischua_2, Ischua_3)
plot_hobos(combined)
}

}
